remove waiting room
