+++ packages/app-extension/src/components/Onboarding/index.tsx
@@ -19,7 +19,6 @@ import { Box, IconButton, ListItemText, Toolbar } from "@mui/material";
 
 import { WithContaineredDrawer } from "../common/Layout/Drawer";
 import { NAV_BAR_HEIGHT } from "../common/Layout/Nav";
-import WaitingRoom from "../common/WaitingRoom";
 
 import { OnboardAccount } from "./pages/OnboardAccount";
 import { RecoverAccount } from "./pages/RecoverAccount";
@@ -31,9 +30,7 @@ export const Onboarding = ({
 }) => {
   const containerRef = useRef();
   const [menuOpen, setMenuOpen] = useState(false);
-  const [action, setAction] = useState<"onboard" | "recover" | "waiting">(
-    "onboard"
-  );
+  const [action, setAction] = useState<"onboard" | "recover">("onboard");
 
   const _ks = useKeyringStoreState();
   const isOnboarded =
@@ -69,12 +66,10 @@ export const Onboarding = ({
         <OnboardingProvider>
           <OnboardAccount
             onRecover={() => setAction("recover")}
-            onWaiting={() => setAction("waiting")}
             {...defaultProps}
           />
         </OnboardingProvider>
       ) : null}
-      {action === "waiting" ? <WaitingRoom /> : null}
       {action === "recover" ? (
         <OnboardingProvider>
           <RecoverAccount
